* internal/develop_rel_1_1_13:
  Update core version to 1.1.13
  Add TC_iot_noti_sub_cb_rate_limit_reached_success Test Case
  Add TC_iot_cap_sub_cb_success Test Case
  add log group base id for dump log
  add dump log function and dump log id list
  bsp: esp: Do not turn off the wifi if network access failed
  misc: use size_t instead of int for length
  iot_log_file: Added new 'log_size' arg on iot_log_file_store()
  test: fix build error according to function name changes
  header define change
  the name change to additional func for easysetup
  test: Add TC for iot_misc_info_load()/iot_misc_info_store()
  test: add TC for _check_prov_data_validation
  test: add TC for _do_status_report

+++ src/iot_util.c
@@ -233,7 +233,7 @@ iot_error_t iot_util_convert_str_uuid(const char* str, struct iot_uuid* uuid)
 	return IOT_ERROR_NONE;
 }
 
-iot_error_t iot_util_convert_uuid_str(struct iot_uuid* uuid, char* str, int max_sz)
+iot_error_t iot_util_convert_uuid_str(struct iot_uuid* uuid, char* str, size_t max_sz)
 {
 	char* ref_id = "42365732-c6db-4bc9-8945-2a7ca10d6f23";
 	int i, written = 0, wrt;
